Add None to the variant.alleles for missing data.
Also port test cases for genotypes into new file, test_genotypes.
